requirements.txt
openai
python-dotenv
autopep8
transformers